chore: fix data loading of activity types
