+++ apps/olat-api/src/services.ts
@@ -4,9 +4,11 @@ import {
   PracticeQuiz,
   PrismaClient,
 } from '@klicker-uzh/prisma'
+import fs from 'fs/promises'
+import path from 'path'
 import { pick } from 'remeda'
-import { activityTypesAvailable } from './static.js'
-import { ActivityOlatConfigurationKey } from './types.js'
+import { fileURLToPath } from 'url'
+import { ActivityOlatConfigurationKey, ActivityType } from './types.js'
 
 const prisma = new PrismaClient()
 
@@ -43,9 +45,25 @@ export async function getCourses(provider: string, providerAccountId: string) {
   }))
 }
 
+async function loadActivityTypes(): Promise<ActivityType[]> {
+  try {
+    const __filename = fileURLToPath(import.meta.url)
+    const __dirname = path.dirname(__filename)
+    const dataPath = path.join(__dirname, '../static/activityTypes.json')
+    const data = await fs.readFile(dataPath, 'utf-8')
+    const activityTypes: ActivityType[] = JSON.parse(data)
+    return activityTypes
+  } catch (error) {
+    console.error('Error reading data:', error)
+    process.exit(1)
+  }
+}
+
 export async function getActivityTypes() {
+  const activityTypes = await loadActivityTypes()
+
   // filter out fields
-  return activityTypesAvailable.map((activityType) => ({
+  return activityTypes.map((activityType) => ({
     ...pick(activityType, [
       'id',
       'path',
@@ -60,6 +78,7 @@ export async function getCourseActivityTypes(
   providerAccountId: string,
   courseID: string
 ) {
+  const activityTypes = await loadActivityTypes()
   const account = await prisma.account.findUnique({
     where: {
       provider_providerAccountId: {
@@ -108,11 +127,11 @@ export async function getCourseActivityTypes(
   }
   const activityKeysGamification = ['course-leaderboard']
 
-  const activityTypes = activityTypesAvailable.flatMap(
+  const availableActivityTypes = activityTypes.flatMap(
     ({ id, title, olatConfigurationKey, isSubselectionRequired }) => {
       // Overview activities: show count in title
       if (olatConfigurationKey in mapOverview) {
-        const count = mapOverview[olatConfigurationKey]?.length || 0
+        const count = mapOverview[olatConfigurationKey]!.length || 0
         return {
           id,
           title: `${title} (${count})`,
@@ -123,7 +142,7 @@ export async function getCourseActivityTypes(
 
       // Subselection activities: only include if they have items
       if (olatConfigurationKey in mapSubselection) {
-        const count = mapSubselection[olatConfigurationKey]?.length || 0
+        const count = mapSubselection[olatConfigurationKey]!.length || 0
         return count > 0
           ? {
               id,
@@ -156,7 +175,7 @@ export async function getCourseActivityTypes(
     }
   )
 
-  return activityTypes
+  return availableActivityTypes
 }
 
 export async function getActivities(
